transform only if finite dimensional columns
