+++ Project.toml
@@ -1,6 +1,6 @@
 name = "ContinuumArrays"
 uuid = "7ae1f121-cc2c-504b-ac30-9b923412ae5c"
-version = "0.9"
+version = "0.9.0"
 
 [deps]
 ArrayLayouts = "4c555306-a7a7-4459-81d9-ec55ddd5c99a"
@@ -24,7 +24,7 @@ FillArrays = "0.11, 0.12"
 InfiniteArrays = "0.10, 0.11"
 Infinities = "0.1"
 IntervalSets = "0.5"
-LazyArrays = "0.21"
+LazyArrays = "0.21.13"
 QuasiArrays = "0.6, 0.7"
 RecipesBase = "1.0"
 StaticArrays = "1.0"
